+++ src/utils/discord_type_guards.ts
